+++ ndk-cache-nostr/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Patch Changes
 
 - Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.2
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
     - @nostr-dev-kit/ndk@2.14.1
 
 ## 0.1.14