+++ ndk-svelte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nostr-dev-kit/ndk-svelte
 
+## 2.4.8
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.2
+
 ## 2.4.7
 
 ### Patch Changes