+++ ndk-test-utils/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Patch Changes
 
 - Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
     - @nostr-dev-kit/ndk@2.14.1
 
 ## 1.0.0